requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
boto3==1.28.84
botocore==1.31.84
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
jmespath==1.0.1
kombu==5.3.5
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
